+++ tests/Rename.test.cpp
@@ -232,4 +232,31 @@ TEST_CASE_FIXTURE(Fixture, "rename_generic_type_correctly_handle_shadowing_3")
     )");
 }
 
+TEST_CASE_FIXTURE(Fixture, "renaming_required_variable_should_also_rename_imported_type_prefixes")
+{
+    auto source = R"(
+        local Types = require("path/to/types")
+
+        type Foo = Types.Foo
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::RenameParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 14};
+    params.newName = "ActualTypes";
+
+    auto result = workspace.rename(params);
+    REQUIRE(result);
+    REQUIRE(result->changes.size() == 1);
+
+    auto documentEdits = result->changes.begin()->second;
+    CHECK_EQ(applyEdit(source, documentEdits), R"(
+        local ActualTypes = require("path/to/types")
+
+        type Foo = ActualTypes.Foo
+    )");
+}
+
 TEST_SUITE_END();